Custom validator improvements
Refactored the way custom validators are invoked - and it's now much simpler!

All you need to do is pass the fully qualifiued class path and name in the validate() method and it will be automatically run once the main validation has been done. The validation errors are also added to the core data errors.

An interesting side-effect of this approach, is you could chain custom validators! I'm not sure why you would ever want to - but if you make the validate method in your custom validator call self.customValiudator() and pass the next fully qualified class path and name, then it would work . Youd need to extend self.validation_errors with the result. Fascinating!